More rearrangements
* Rename of APPLICATION --> MASK_DEMO to ensure the root class is clear
in intent.
* Removal of the mask_demo.ecf and *.rc files as there is no need for
them--the target now lives in the mask.ecf file without issue.

NOTE: This also means that demo classes from the demo folder (cluster)
are available to the library reuse consumer for use in their projects,
which might create an unwanted dependency from the library
creator/designers point of view. However, this is in response to what
appears to be a bug in the parser and compiler for Eiffel Studio and
Eiffel. Specifically:

If I have a primary library target with no root class and feature and
<exclude>/my_folder$</exclude> in that target; a subsequent and
subordinate target cannot bring back the excluded "my_folder" by simply
<include>/my_folder$</include>. While the Project tool in Eiffel Studio
sees the inclusion, it will not see the MY_ROOT_CLASS within it, where
the subordinate target has stated <root class="MY_ROOT_CLASS"
feature="make">.