New "_docs" Folders
* NOTE: I am moving the DOC_* classes within their respective clusters
to a "_docs" folder/cluster. Why?

1. Problem: Having the DOC_* class mixed in with the other working
classes, causes the DOC_* to get lost alphabetically among the others.
Placing it within a "_docs" cluster, moves the cluster to the top of the
list and distinguishes it as a container for very particular types of
classes that are NOT expected to be reusable code.

2. See last part of point #1: Problem: DOC_* classes mixed in with other
code have to be constantly "weeded" (parsed, culled, mentally filtered)
when reading the contents of the cluster. To save this labor, the DOC_*
classes are moved to the "_docs" cluster/folder.

The "_docs" folder/cluster seems to be a happy-medium or compromise
between the original layout where all DOC_* classes were in a
"documentation" cluster in the root. I can concede to the "_docs" being
in each cluster, which clearly ID's that documentation as belonging to
the cluster. However, as stated: I am moving the DOC_* classes into the
"_docs" sub-cluster to disambiguate them from their actual and reusable
library classes.